+++ src/FootballDataApi/Models/Winner.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 
-namespace FootballDataApi.Models.Competitions;
+namespace FootballDataApi.Models;
 
-public sealed record Winner
+public class Winner
 {
     public int Id { get; set; }
 
@@ -16,7 +16,11 @@ public sealed record Winner
 
     public string Address { get; set; }
 
+    public string Phone { get; set; }
+
     public string Website { get; set; }
+
+    public string Email { get; set; }
 
     public int Founded { get; set; }
 
@@ -26,4 +30,5 @@ public sealed record Winner
 
     public DateTime LastUpdated { get; set; }
 }
+
